implement account signup page.
make errmsg tag use ActionMessages.GLOBAL_MESSAGE form property if no property
attribute is specified. this requires the use of the bind tag from the
unstandard taglib from jakarta.

add establishSecurityContext() to the CosmoSecurityManager interface to allow
cosmo components to programatically log in a user (as opposed to relying on
Acegi Security or JAAS to do it for us). this makes CosmoSecurityManagerImpl
depend on AuthenticationManager, which is a circular dependency, so remove the
loadUser method from CosmoSecurityManager and allow our Authentication dao to
access our User dao directly.

support anonymous authentication and run-as authorization with
CosmoSecurityContext. add isAnonymous method to the interface and
AUTHORITY_RUN_AS_ROOT constant to the implementation class. we only support
running as root for the moment. this allows the user interface to identify
itself to jackrabbit as a root user in order to create the home directory for
a user during account signup (who is known to the ui as an anonymous user).

fix the "username exists" error message set by UserAction so that it shows up
next to the username form field correctly.

change "home page" back to "welcome page" to not conflict with the notion of
the user's home directory.

unify acegi security filter security interceptor configurations for web ui and
dav. any url not explicitly configured is accessible only to root users.

fix urlrewrite rules to allow all legal username characters.


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@22 5fd0b11a-f2f7-0310-948e-c8f80ebed640