get ticket authentication mostly working. there's some wonkiness when using cadaver and firefox, but straight up curl from the command line works correctly. tickets aren't inherited yet so you only have access to the specific ticketed resource.
refactor jcr/ticket mapping stuff into JCRUtils class. add TicketDao and 
JCRTicketDao that can retrieve a ticket for a particular node from the 
repository. this is only different from the code in CosmoDavResourceImpl in 
that the latter already has access to the ticketed resource's node. 

revert back to SimpleAccessManager for repository security. rather than having
the repository enforce permissions at the item level, we should treat it more
like a typical webapp-backing sql db and have a single repository user with
permissions on the entire repository (or maybe per-workspace). 


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@364 5fd0b11a-f2f7-0310-948e-c8f80ebed640