define the mix:ticketable mixin node type that allows nodes to have child ticket nodes and add the mixin type to the nt:folder and nt:file resources created through webdav.
this approach requires a mix:ticketable node to have 0..n child nodes named
"dav:ticket" - ticket id and node name are no longer the same thing. ticket:id
is a new property on the ticket:ticket node type. sync the cosmo apis to this
change.

implement ticket inheritance such that if a ticket is not found on the
directly addressed node, its ancestors are searched to find the ticket. this
allows me to MKTICKET on /hi and then use that ticket to access
/hi/project.xml.

fix namespaces in ticket response xml that got busted with an erroneous search
and replace.


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@366 5fd0b11a-f2f7-0310-948e-c8f80ebed640