take the revolutionary step of removing CalendarDao altogether. the reasoning:
1) the plumbing within cosmo to pass the spring application context
down through the layers so that the import/export handlers could use
them to retrieve the daos made several classes needlessly complex

2) CalendarDao's apis took a path argument instead of a Node. this
meant we had two separate Node instances on either side of the dao
boundary that needed to be synced. inefficient and overly complex.

3) JcrCalendarDao was nothing more than an extremely thin wrapper
around JcrCalendarMapper anyway. the real logic of adding node types
and so forth lives in the latter class. it has only static methods so
there's no need to pass an instance between webdav objects.

this change simplifies a lot of things with really no cost.


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@817 5fd0b11a-f2f7-0310-948e-c8f80ebed640