allow the client to specify many tickets in the request, in both headers and parameters, and use the first one we find that is valid (eg that exists on the resource or one of its ancestors).
change the way we calculate the path for a ticket token to take into
account that the resource path in a cosmo dav request is always the
path info of the request uri. this removes the need for the
resourcePathPrefix attribute of TicketProcessingFilter.

fix JcrTicketDao.getTicket to actually return null when the ticket is
not found rather than throwing an exception.


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@1389 5fd0b11a-f2f7-0310-948e-c8f80ebed640