add the beginnings of a generic repository migration framework.
the entry point is o.o.c.m.Migrator. this class uses command line
arguments to configure instances of MigratorClient for a source
repository (where the data is being migrated from) and a target
repository (where the data is being migrated to).

a "migration" is a tool that copies the data from the source
repository into the target repository, converting it into the format
expected by the target repository. a separate migration is created for
each application version and can convert either from the previous
version to the current one (an "up" migration), or from the current
one to the previous (a "down" migration).

eventually Migrator will be able to chain migrations together, so that
one could migrate a repository through several application versions
ine one go.

the idea is that an application will use a tool like the maven javaapp
plugin to build a single executable jar containing the framework jar
and the application-specific migration classes, so that a migration
can be as easy as this:

  java -jar application-migrator.jar -sourceconfig ... -sourcepath ... \
                                     -targetconfig ... -targetpath ...



git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@1704 5fd0b11a-f2f7-0310-948e-c8f80ebed640