Huge checkin.
It's after midnight, but I have stress tests working.

I implemented the option of running recurring test cases, it took a lot of plumbing in TestObject and some code additons to the tests I wrote to work with it.

Right now only cosmo_basicquery and cosmo_timerangequery implement recurring tests which run a specified amount of time after runing testStart.

I wrote and added stress.py which is the beginning of the stress test library
Currently stress.py can take an arbitrary list of modules.classes and and arbitrary amount of threads and run the modules.classes.fullRun() in each thread, then calculate all the passes and fails.

I had to write even more plumbing in to TestObject in order to get useful output when failures happen when running test objects in threads.

Lots more to do, but cosmo_fullsuite still runs as before, if recurrence=int is specified then only the tests that have implemented it will run it.


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@1798 5fd0b11a-f2f7-0310-948e-c8f80ebed640