Migration03 now adds a /cosmo:system node with a cosmo:schema child node for storing properties related to the cosmo schema. the only property currently stored is cosmo:schemaVersion.
before doing anything else, Migration03 checks to see if the schema
version has been written into the repository. it bombs out if there is
a version but it is not 0.3. if the schema version has not been
written, then it assumes this is the first attempt at migration and
registers namespaces and node types.


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@1826 5fd0b11a-f2f7-0310-948e-c8f80ebed640