put json requests in the security filter path.
This commit may be controversial, since I tried to please everyone. I use HttpSessionIntegrationFilter 
(as opposed to an HttpRequestIntegrationFilter) so that if there already is a session with security credentials
cached, those will be used. Otherwise, Basic auth or anon can be used. 

So you can use session or basic auth. 

git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@2204 5fd0b11a-f2f7-0310-948e-c8f80ebed640