bug: 5958
Upgrade to acegi-security 1.0.3

No need to subclass AnonymousAuthenticationFilter, as things have been
refactored in the latest acegi src.  For one, the method that NoSessionAnonymouAuthenticationFilter
overrides is no longer present in the super class.  Also, AnonymousAuthenticationFilter
has been refactored to use a AuthenticationDetailsSource to generate a details object.
The default implementation creates a WebAuthenticationDetails object, which has been
refactored to always call request.getSession(false), which is what
NoSessionAnonymousAuthenticationFilter was created to do (don't force a session create).

git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@3051 5fd0b11a-f2f7-0310-948e-c8f80ebed640