bug: 7309
Fix CaldavReport.doQuery() to continue recursive execution if colleciton
is not a calendar.  This prevented freebusy rollups from working because 
the report was run against the home collection, which is not a calendar, 
and the report was exiting without evaluating the child calendars.

Also fix a bug in InstanceList where an invalid instance was being added
for recurring events.

Add test to verify doQuery is being called recursively.

git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@3104 5fd0b11a-f2f7-0310-948e-c8f80ebed640