Test to demonstrate speed of two "get*" calls vs one "get*" and a variable creation/assignment.
In firefox, it appears that one get* and var creation is about 3 times faster. This indicates that
all situations requiring two Model attribute accesses in the same scope should be handled
with the creation of a local variable. This was the intuitive result, I believe, but this test provides
good evidence.

git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@4428 5fd0b11a-f2f7-0310-948e-c8f80ebed640