switch to including relative URLs (path-absolute) in dav responses. this requires new implementations of StandardResourceLocator and StandardResourceLocatorFactory that are based on URL and URI rather than ServiceLocator.
DAV:alternate-URI-set no longer includes any URIs, because we don't have any
apis that are capable of generating URIs from all different protocols other
than ServiceLocator, which as of this commit is no longer used in the dav
subsystem. this is no big loss. someday when we get a pan-server
ResourceLocator api we'll be able to do it.

happily the new resource locator stuff simplifies StandardDavRequest's
construction of request and destination locators. another side effect is that
we're now encoding and decoding dav URLs correctly (need to propagate this
to atom and cmp).


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@5697 5fd0b11a-f2f7-0310-948e-c8f80ebed640