Merge r6141 from 0.9 to trunk. Original commit message:
Render ui even if there are no collections to allow users to create a new collection if they have 0 collections.

Note: this leads to a weird state where the calendar ui is available, but interacting with it causes errors. Since these errors do not have serious consequences (they just look silly) I'm going to open a bug and let bug council decide whether this should actually be fixed in 0.9.



git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@6146 5fd0b11a-f2f7-0310-948e-c8f80ebed640