fix for bug 11336 - triaging LATER items to DONE doesn't work the first time.
The problem was that the setTriage method that was automagically created for the 
class was behaving like all the other setXXX methods in that it doesn't add a new
property to the modification object if the value to be set is the same as the master.  

In the case of triage status though, we ignore the master, so we always want to be setting stuff on the modification object when changing triageStatus.



git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@6265 5fd0b11a-f2f7-0310-948e-c8f80ebed640