re-factoring error handling code in DetailFormConverter.
This is preliminary work for bug 11377.

Basically, the formatting of the errors are handled in one place, and errors are represented 
as objects instead of strings. 

The whole reason this is being done (besides being more flexible in the long term) is that 
I need to know when there is a duplicate error, and not to add it (for example, the timezone 
region error)


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@6282 5fd0b11a-f2f7-0310-948e-c8f80ebed640