Major Javascript build rework.
All Javascript source can now be found under /cosmo-js, and is put in place during
mvn war:inplace or mvn package in /cosmo.

I've added a profile to /cosmo/pom.xml to allow for the old dev workflow, in which
the javascript exposed via tomcat and used in the application is under version control.
This can now be accomplished by doing

mvn -Pjs-dev compile war:inplace

in /cosmo, with the caveat that sym linking must be enabled in Tomcat. In the server
bundle, this can be done by editing SNARF_DIST_HOME/tomcat/conf/Catalina/localhost/chandler.xml

and adding

allowLinking="true"

to the parameters to the Context element, like:

<Context reloadable="false"
         docBase="/path/to/war"
         allowLinking="true">


git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@6803 5fd0b11a-f2f7-0310-948e-c8f80ebed640